requirements.txt
# Core Flask Dependencies
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==22.0.0

# Security Dependencies
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
bleach==6.0.0

# Google Cloud Dependencies
google-cloud-firestore==2.11.0
functions-framework==3.4.0

# HTTP and API Dependencies
requests==2.31.0

# Environment and Configuration
python-dotenv==1.0.0

# Testing Dependencies
pytest==7.4.0

# Email and Notifications
# Note: smtplib is included in Python standard library

# Data Analysis and Statistics
# Note: statistics is included in Python standard library

# Development and Debugging
Jinja2==3.1.2
MarkupSafe==2.1.3

# Optional: For enhanced features
# Uncomment if you want additional analytics capabilities
# pandas==2.0.3
# numpy==1.24.3
# scikit-learn==1.3.0